smart_research_assistant_project/requirements.txt
# Core packages
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.10
langgraph>=0.0.20
langsmith>=0.0.69
langchain-text-splitters>=0.0.1

openai>=1.3.0

# API serving
fastapi>=0.110.0
uvicorn>=0.27.0
langserve>=0.0.40
sse-starlette>=2.0.0

# Vector databases
chromadb>=0.4.18

# Document loading and processing
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Environment variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.2
typing-extensions>=4.8.0
tqdm>=4.66.1

